+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## unreleased changes
+* `BundledAssetConfiguration` に `BundledTextAssetConfiguration` を追加
+
 ## 2.4.0
 * `GameConfiguration` に `assetBundle` を追加
 * `AssetBundleConfiguration` および `BundledAssetConfiguration` を追加